aoc2024: Optimize Day 6, part two -- Vec -> HashSet.
Since the only thing that determines the guard's next move is the
previous move, we don't need to scan a slice of the guards history,
but just a single move. Now we only check if a move has been made and
we store them in a HashSet.

This improved performance 96.5x fold.

cyclecheck | real 19m37,884s
hashset    | real  0m12,198s

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>