aoc2024: Solve Day 13, part two.
Now instead of counting the fences directly in the walk function, we
merely mark where the fences are in a grid (tridimensional, that is, as
each cell can have up to four sides), and then later count the sides by
counting the transitions from not-fenced to fenced. Some modulo trickery
is involved.

Signed-off-by: Vinícius Schütz Piva <vinicius.vsczpv@outlook.com>